After learning bootstrap
I deleted the UL and created a grid in the header with bootstrap. I
modified the background of the header by adding a class=‘header’. I
formed it in CSS also. In the index.html I added class=‘text-centered’
to align the links in the middle of the grid, and I created buttons
from the links.