For sale edited and jumping on page solved
I made the tabs jump to anchors on the pages, so there is now three
anchor according to the three tabs (musthave, forsake, help). And I
linked a spreadsheet into the pages, because it seemed the easiest
solution for creating runtime editable sheets. It doesn’t work
perfectly so far, but later I try to solve it :D