Navbar, text centered, button
I tried to add a navigation fixed-bottom scrollable bar. It’s there but
i think it doesn’t work quite properly yet. If it is hidden, the
original tabs are working properly and jumping on the page.

I made the « edit here » link a nice blue button :)
And I centered the content.