Added map
I added the map to the main page, but I still haven’t figured out how
to add multiple markers. It is sooo JavaScript. But I hope a few
classes later it will be a bit easier.
I added some comments to the code that way it will be easier for you to
follow my changes.