indenting code, working on h1 and image carousel
I am indenting the code so it is easier to read (done until line 85)

I want to change the colour of the h1 heading to the cambridge blue.
Any idea why this isn’t working?

I have been trying to center the images of the carousel but so far I
have been unsuccessful :( I also don’t know how to stop the images from
‘jumping’.